scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9 